WIP: Refactor logic out of handle_messages
I don't like the current solution because you can't open
the modified changes or staged changes if they have nothing in them.

I believe this should be possible, but running the check to switch to
the difftype that has the files in it on every frame just doesn't seem
like the right approach.

It works as intended with logic in handle_messages, and only swaps
on refresh and when you open a folder. You can still click on staged
changes and view them even if their isn't any files there.